GridMethod

Documentation for GridMethod.

This is a project aiming at implementing the Grid Method theoretically developed by Bürgisser, Cucker, Krick, Lairez, Shub and Tonelli-Cueto. The aim of this project is to develop parallelizable algorithms for solving real polynomial systems and computing the topology of real algebraic and semialgebraic sets...

GridMethod.grid_point

GridMethod.grid_point

GridMethod.grid_point

Base.eltype

GridMethod.D

GridMethod.Han_min

GridMethod.Id

GridMethod.K

GridMethod.K

GridMethod.Wnorm

GridMethod.Wnorm_poly

GridMethod.Wnorm_term

GridMethod.Wnorm_term

GridMethod._degrees

GridMethod._exposcoeffs

GridMethod._get_polys

GridMethod._get_vars

GridMethod._isHan

GridMethod._monomial

GridMethod._norm

GridMethod.coord_opt

GridMethod.cube_nthinitgrid

GridMethod.cube_pushsubdivided!

GridMethod.fpoint

GridMethod.point

GridMethod.radius

GridMethod.rand_poly

GridMethod.refine_grid

GridMethod.refine_grid!

GridMethod.step_to_ratio

GridMethod.tree_nextleaves

GridMethod.tree_nthleaves

GridMethod.tree_root

GridMethod.Δm1

GridMethod.grid_point — Type

grid_point{T}

Structure representing a node in the grid.

Fields

x: Vector of coordinates of the point.

fx: Value of the function at x.

step: Step of the subdivided point.

GridMethod.grid_point — Method

grid_point(x, fx, step)

Creates a new grid_point object with the given coordinates, function value, and step. Promotes the eltype and type of x and fx resp. Also ensures step is Int.

GridMethod.grid_point — Method

grid_point(fun::Function, x, step)

Creates a new grid_point object with the given function, coordinates, and step, where the function value is computed by calling fun(x...).

Base.eltype — Method

eltype(::grid_point{T}) where {T}

Returns the type of the elements in the x field of a grid_point object.

GridMethod.D — Function

D(Jx, x, n = length(x))

Returns the matrix Jx multiplied by the difference between the identity matrix and the outer product of x and its conjugate transpose.

Keyword arguments:

n: The size of the identity matrix.

GridMethod.Han_min — Method

Han_min(H, L, C)

Returns the minimum function value in H scaled by a factor of γ, where γ = 1 - inv(C) * L.

GridMethod.Id — Method

Id(::Type{T}, n) where T

Returns the identity matrix of size n and type T.

GridMethod.K — Function

K(F; compiled = true)

Returns a function that calculates the condition number of a system of polynomials F.

Keyword arguments:

compiled: A boolean indicating whether the system should be compiled before being passed to the condition number function.

GridMethod.K — Method

K(Wnorm, Δm1)

Returns a function that calculates the condition number of a system of polynomials given its Weyl norm and inverse degree matrix.

GridMethod.Wnorm — Method

Wnorm(F; vars = _get_vars(F), expanded = true)

Returns the Weyl norm of a system of polynomials F.

Keyword arguments:

vars: The variables of the system.

expanded: A boolean indicating whether the polynomials should be expanded before calculating the Weyl norm.

GridMethod.Wnorm_poly — Method

Wnorm_poly(poly; vars, expanded = false)

Returns the Weyl norm of a polynomial poly.

Keyword arguments:

vars: The variables of the polynomial.

expanded: A boolean indicating whether the polynomial should be expanded before calculating the Weyl norm.

GridMethod.Wnorm_term — Method

Wnorm_term(I, C)

Returns the term of the Weyl norm corresponding to the exponents I and coefficient C.

GridMethod.Wnorm_term — Method

Wnorm_term(IC) = Wnorm_term(IC...)

GridMethod._degrees — Function

_degrees(F)

Returns the degrees of a system of polynomials F.

GridMethod._exposcoeffs — Function

_exposcoeffs(poly; vars, expanded = false)

Returns a list of tuples containing the exponents and coefficients of a polynomial poly.

Keyword arguments:

vars: The variables of the polynomial.

expanded: A boolean indicating whether the polynomial should be expanded before extracting the exponents and coefficients.

GridMethod._get_polys — Function

_get_polys(F)

Returns the polynomials of a system F.

GridMethod._get_vars — Method

_get_vars(F)

Returns the variables of a system of polynomials F.

GridMethod._isHan — Method

_isHan(fx, step, C)

Returns true if C * step_to_ratio(step) < fx, and false otherwise.

GridMethod._monomial — Function

_monomial(X, I, n = length(X))

Returns the monomial of variables X with exponents I.

Keyword arguments:

n: The number of variables.

GridMethod._norm — Method

_norm(v)

Returns the Euclidean norm of a vector v.

GridMethod.coord_opt — Method

coord_opt(::Type{T}, dim) where {T} = ntuple(_ -> (one(T), -one(T)), dim)

Returns an iterator which generates all possible dim-dimensional vectors that have all components either one or -one. The eltype of the vector is T, which defaults to Float64.

GridMethod.cube_nthinitgrid — Method

cube_nthinitgrid(T=Float64, fun, dim, n; ratio = 0.5, root = tree_root(T, dim),
                 coord_opt = coord_opt(T, dim))

Returns a list of grid_point objects that form an initial grid with 2^n points in each dimension.

Arguments:

T (optional): Numeric type to represent the coordinates of the points.

fun: The function to be evaluated at each point in the grid.

dim: The dimension of the grid.

n: The number of times the grid should be recursively divided.

Keyword arguments:

ratio: The ratio between the length of a branch and of the previous.

root: The root of the grid (a point in dim-dimensional space).

coord_opt: The possible directions in which the next points can be generated.

GridMethod.cube_pushsubdivided! — Method

cube_pushsubdivided!(G, p, step, fun; dim = length(p),
                     coord_opt = coord_opt(eltype(p), dim))

Pushes the points obtained by subdividing the region around p into G.

Arguments:

G: A list of grid_point objects, where to push the new points.

p: A point in dim-dimensional space.

step: The step at which the point was added to the grid.

fun: The function to be evaluated at each point.

Keyword arguments:

dim: The dimension of the grid.

coord_opt: The possible directions in which the next points can be generated.

GridMethod.fpoint — Method

fpoint(g::grid_point)

Returns the function value of the grid_point object g.

GridMethod.point — Method

point(g::grid_point)

Returns the coordinates of the grid_point object g.

GridMethod.radius — Method

radius(g::grid_point)

Returns the radius of the grid_point object g, which is computed as 1/2^i, where i is the step at which the point was added to the grid.

GridMethod.rand_poly — Method

rand_poly(::Type{T}=Float64, X, d; coeffs = I -> randn(T), n = length(X))

Returns a random polynomial with coefficients of type T in variables X with degree d.

Keyword arguments:

coeffs: A function that takes a tuple of exponents and returns a coefficient.

n: The number of variables.

GridMethod.refine_grid! — Method

refine_grid!(G, fun, C, m, dim; isfine = _isHan,
             pushsubdivide! = cube_pushsubdivided!)

Modifies the input list G in place to form a refined grid satisfying the Han condition.

Arguments:

G: A list of grid_point objects.

fun: The function to be evaluated at each point in the grid.

C: A parameter used in the Han condition.

m: A parameter used to determine when to stop refining the grid.

dim: The dimension of the grid.

Keyword arguments:

isfine: A function that determines whether a point satisfies the Han condition.

pushsubdivide!: A function that pushes the points obtained by subdividing a region into a list of points.

GridMethod.refine_grid — Method

refine_grid(T=Float64, fun, C, m, dim; step₀ = Int(ceil(log2(C))),
            G = cube_nthinitgrid(T, fun, dim, step₀), isfine = _isHan,
            pushsubdivided! = cube_pushsubdivided!)

Returns a list of grid_point objects that form a refined grid satisfying the Han condition.

Arguments:

T (optional): Numeric type to represent the coordinates of the points.

fun: The function to be evaluated at each point in the grid.

C: A parameter used in the Han condition.

m: A parameter used to determine when to stop refining the grid.

dim: The dimension of the grid.

Keyword arguments:

step₀: The initial step of the grid.

G: The initial grid.

isfine: A function that determines whether a point satisfies the Han condition.

pushsubdivide!: A function that pushes the points obtained by subdividing a region into a list of points.

GridMethod.step_to_ratio — Method

step_to_ratio(step) = inv(exp2(step))

Converts the subdivided index step of a grid_point into the ratio of such point.

GridMethod.tree_nextleaves — Method

tree_nextleaves(p, ratio; dim = length(p), coord_opt = coord_opt(eltype(p), dim))

Returns an iterator which generates the next leaves of the tree starting from p with a length ratio of r.

Keyword arguments:

dim: The dimension of the tree.

coord_opt: The possible directions in which the next leaves can be generated.

GridMethod.tree_nthleaves — Method

tree_nthleaves(dim, n; ratio = 0.5, root = tree_root(dim),
               coord_opt = coord_opt(dim))

Computes the nth leaves of a 2^dim-branched tree in a dim-dimensional space, where dim is the dimension of the tree (i.e. 2 for a 2D tree, 3 for a 3D tree), and n is the number of times the tree should be recursively branched.

Keyword arguments:

ratio: The ratio between the length of a branch and of the previous.

root: The root of the tree (a point in dim-dimensional space).

coord_opt: The possible directions in which the next leaves can be generated.

GridMethod.tree_root — Method

tree_root(::Type{T}, dim) where {T} = [zeros(T, dim)]

Returns an array containing a single zero-vector of length dim and type T, which defaults to Float64. It is meant to be used as the root node of a tree fractal.

GridMethod.Δm1 — Method

Δm1(F)

Returns the inverse of the degree matrix of a system of polynomials F.