Cross-Tree fractal

The nodes of the grid form the leaves of a cross-formed tree fractal.

GridMethod.pmones — Method

pmones(::Type{T}=Float64, dim)

Returns an iterator which generates all possible dim-dimensional vectors that have all components either one or -one.

GridMethod.tree_nextleaves — Method

tree_nextleaves(p, ratio; dim = length(p), dirs = pmones(eltype(p), dim))

Returns an iterator which generates the next leaves of the tree starting from p with a length ratio of r.

Keyword arguments:

dim: The dimension of the tree.

dirs: The directions in which the next leaves will be generated.

GridMethod.tree_nthleaves — Method

tree_nthleaves(dim, n; ratio = 0.5, root = tree_root(dim),
               dirs = pmones(eltype(p), dim))

Computes the nth leaves of a 2^dim-branched tree in a dim-dimensional space, where dim is the dimension of the tree (i.e. 2 for a 2D tree, 3 for a 3D tree), and n is the number of times the tree should be recursively branched.

Keyword arguments:

ratio: The ratio between the length of a branch and of the previous.

root: The root of the tree (a point in dim-dimensional space).

dirs: The possible directions in which the next leaves can be generated.

GridMethod.tree_root — Method

tree_root(::Type{T}=Float64, dim)

Returns an array containing a single zero-vector of length dim and type T. It is meant to be the root node of a tree fractal.